requirements.txt
crewai>=0.86.0
crewai-tools>=0.17.0
streamlit>=1.40.0
pydantic>=2.0.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-ollama>=0.2.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
sqlparse>=0.5.0
pyyaml>=6.0.0
